I changed some categories that were redundant, or categories with only a single entry.  Example: chants was a category that only had one entry, so I removed that category mark.
There were three separate categories: general,information,general information
and I merged them into one: general information.

Things of that sort.

Anywhere where there was a line of --- bars, but it was broken with
a space in the middle, I fixed it.
Anywhere where there was a line of === bars, but it was broken with
a space in the middle, I fixed it.

removed some spaces/tabs at end of lines, where it could not be seen, but
could cause problems with my converter program.

Mostly compressed the large, unkempt categroy mess into more standardized
categories.